projet/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
numpy>=1.21.0
opencv-python>=4.8.0
pandas>=2.0.0
tabulate>=0.9.0
python-dateutil>=2.8.2
pytz>=2023.3
scikit-learn>=1.3.0
scikit-image>=0.21.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
gunicorn>=21.0.0
drf-yasg>=1.21.0
